Backend/requirements.txt
alembic==1.12.1
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Injector==0.15.0
Flask-Migrate==4.0.5
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
importlib-resources==6.1.0
injector==0.21.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mistune==3.0.2
packaging==23.2
#psycopg2-binary==2.9.9 TODO: Install it manually
PyJWT==2.8.0
pyrsistent==0.20.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
#rpds-py==0.10.6 TODO: Install it manually
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==2.3.7
